requirements.txt
telethon>=1.34.0
cryptg>=0.4.0
rich>=13.0.0
filelock>=3.12.0
Pillow>=10.0.0
aiofiles>=23.0.0
python-socks>=2.4.0
tqdm>=4.66.0
